+++ processing/CHANGELOG.md
@@ -2,19 +2,21 @@
 
 This is an manual and incomplete list of change to the data structure.
 
-## 2025-09-11
+## 2025-09-11, -16
 
 ### `bikelanes`
 
 - Add attributes `operator_type`, `informal`, `covered`
 - Include ways that are privately operated, indoor or informal
 - Include ways that are `access=destination` in the data
+- Include ways on `highway=service` when also bike tagging present
 - Categorize bicycle roads with "Kfz frei" as `bicycleRoad_vehicleDestination`
 
 ### `roads`, `roadsPathClasses`,
 
 - Add attributes `operator_type`, `informal`, `covered`
 - Include ways that are privately operated, informal
+- Include ways on `highway=service` when also bike tagging present
 
 ## 2025-07-29